# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Darbby Platform API — Backend for customer and merchant apps
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List my vehicle profiles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VehicleProfile"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Add a vehicle profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleProfile"
  /vehicles/{id}:
    put:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update a vehicle profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleProfile"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete a vehicle profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "204":
          description: Deleted

  # ─── TRIPS ────────────────────────────────────────────────────────────────
...
  /geo/route:
    post:
      operationId: getRoute
      tags: [geo]
      summary: Get route polyline and distance from Google Maps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
      responses:
        "200":
          description: Route information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"

  # ─── MERCHANT STATS ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    CreateVehicleRequest:
      type: object
      properties:
        nickname:
          type: string
        vehicleType:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        fuelType:
          type: string
          enum: [PETROL_91, PETROL_95, DIESEL, ELECTRIC, HYBRID]
        plateNo:
          type: string
        tankCapacityLiters:
          type: number
        isPrimary:
          type: boolean
      required: [nickname, vehicleType, make, model, year, fuelType]
    VehicleProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        nickname:
          type: string
        vehicleType:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        fuelType:
          type: string
        plateNo:
          type: string
        tankCapacityLiters:
          type: number
        isPrimary:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, nickname, vehicleType, make, model, year, fuelType]

    # ─── Trip ────────────────────────────────────────────────────────────────
...
    RouteRequest:
      type: object
      properties:
        originLat:
          type: number
        originLng:
          type: number
        destLat:
          type: number
        destLng:
          type: number
      required: [originLat, originLng, destLat, destLng]
    RouteResponse:
      type: object
      properties:
        polyline:
          type: string
        distanceMeters:
          type: integer
        durationSeconds:
          type: integer
        distanceText:
          type: string
        durationText:
          type: string
      required: [polyline, distanceMeters, durationSeconds]

    # ─── Stats ────────────────────────────────────────────────────────────────
...